requirements.txt
absl-py==1.0.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.11.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.12.0
CacheControl==0.14.4
cachetools==5.0.0
certifi==2021.10.8
cffi==2.0.0
charset-normalizer==2.0.12
click==8.1.3
cmake==4.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.13
channels==4.3.1
channels-redis==4.3.0
django-js-asset==3.1.2
django-mptt==0.18.0
exceptiongroup==1.3.0
filetype==1.2.0
firebase_admin==7.1.0
Flask==2.2.3
flask-cors==6.0.1
flatbuffers==2.0
fonttools==4.62.0
future==1.0.0
gast==0.5.3
geocoder==1.38.1
google-api-core==2.28.1
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gpt4all==2.8.2
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
h5py==3.6.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.3
ImageIO==2.37.3
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.5.0
lazy-loader==0.5
libclang==13.0.0
llvmlite==0.46.0
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.10.8
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.28
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
oauthlib==3.2.0
onnxruntime==1.23.2
openai==1.99.9
opencv-python==4.5.5.64
opencv-python-headless==4.13.0.92
opt-einsum==3.3.0
packaging==26.0
pandas==1.4.2
pefile==2024.8.26
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
proto-plus==1.26.1
protobuf==3.20.3
psycopg2-binary==2.9.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.2
PyJWT==2.10.1
PyMatting==1.1.15
PyMsgBox==1.0.9
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.9.0.post0
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pywin32-ctypes==0.2.3
qrcode==8.2
ratelim==0.1.6
referencing==0.37.0
rembg==2.0.69
requests==2.27.1
requests-oauthlib==1.3.1
rpds-py==0.30.0
rsa==4.8
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.4
sqlparse==0.5.5
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.2.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2026.1
urllib3==1.26.9
watchdog==6.0.0
Werkzeug==2.2.3
whitenoise==6.12.0
wikipedia==1.4.0
wrapt==1.14.0
